requirements.txt
asgiref==3.8.1
cffi==1.16.0
Django==5.0.7
djangorestframework==3.15.2
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
zope.event==5.0
zope.interface==6.4.post2
